adds fixes to for fixtures
