+++ src/components/fixtures/FixtureHeadToHeadStats.tsx
@@ -20,7 +20,7 @@ export default function FixtureHeadToHeadStats({ fixture, teamActions }: Props) 
             action: mapSportsActionToAthstatName(t.action) as ITeamActionName
         }
     })
-    const taParser = new TeamActionsParser(formatedTeamActions, fixture.team.athstat_id, fixture.opposition_team.athstat_id);
+    const taParser = new TeamActionsParser(formatedTeamActions, fixture?.team?.athstat_id ?? '', fixture?.opposition_team?.athstat_id ?? '');
 
     return (
         <TitledCard title="Head to Head" icon={Shield} >
@@ -29,13 +29,13 @@ export default function FixtureHeadToHeadStats({ fixture, teamActions }: Props) 
 
                 <div className="flex flex-row gap-1" >
                     <div className="flex flex-1 items-center justify-start" >
-                        <TeamLogo className="w-6 h-6" teamName={fixture.team.athstat_name} url={fixture.team.image_url} />
+                        <TeamLogo className="w-6 h-6" teamName={fixture?.team?.athstat_name} url={fixture?.team?.image_url} />
                     </div>
                     <div className="flex flex-[3] items-center justify-center text-center " >
                         <p>Team Stats</p>
                     </div>
                     <div className="flex flex-1 items-center justify-end" >
-                        <TeamLogo className="w-6 h-6" teamName={fixture.team.athstat_name} url={fixture.opposition_team.image_url} />
+                        <TeamLogo className="w-6 h-6" teamName={fixture?.team?.athstat_name} url={fixture?.opposition_team?.image_url} />
                     </div>
                 </div>
