+++ src/components/fixtures/FixtureRosters.tsx
@@ -37,13 +37,13 @@ export default function FixtureRosters({ fixture }: Props) {
 
     const tabs: TabViewHeaderItem[] = [
         {
-            label: fixture.team.athstat_name,
+            label: fixture?.team?.athstat_name,
             tabKey: "home-team",
             className: "flex-1 w-1/2"
         },
 
         {
-            label: fixture.opposition_team.athstat_name,
+            label: fixture?.opposition_team?.athstat_name,
             tabKey: "away-team",
             className: "flex-1 w-1/2"
         }
@@ -52,12 +52,12 @@ export default function FixtureRosters({ fixture }: Props) {
     const rosters = fetchedRosters ?? [];
 
     const awayRoster = rosters.filter((r) => {
-        return r.team_id === fixture.opposition_team.athstat_id;
+        return r.team_id === fixture?.opposition_team?.athstat_id;
     });
 
 
     const homeRoster = rosters.filter((r) => {
-        return r.team_id === fixture.team.athstat_id;
+        return r.team_id === fixture?.team?.athstat_id;
     })
 
     const handleClickPlayer = (player: IProAthlete) => {
