+++ src/screens/FixtureScreen.tsx
@@ -1,10 +1,6 @@
-import { useNavigate, useParams } from 'react-router-dom';
+import { useParams } from 'react-router-dom';
 import { IFixture } from '../types/games';
-import TeamLogo from '../components/team/TeamLogo';
-import { format } from 'date-fns';
-import { fixtureSumary, summerizeGameStatus } from '../utils/fixtureUtils';
-import { Minus } from 'lucide-react';
-import { ArrowLeft } from 'lucide-react';
+import { fixtureSumary } from '../utils/fixtureUtils';
 import FixtureScreenOverview from '../components/fixtures/FixtureScreenOverview';
 import useSWR from 'swr';
 import { gamesService } from '../services/gamesService';
@@ -14,7 +10,6 @@ import { boxScoreService } from '../services/boxScoreService';
 import { FixtureScreenHeader } from '../components/fixtures/FixtureScreenHeader';
 import TabView, { TabViewHeaderItem, TabViewPage } from '../components/shared/tabs/TabView';
 import FixtureHeadToHeadStats from '../components/fixtures/FixtureHeadToHeadStats';
-import BlueGradientCard from '../components/shared/BlueGradientCard';
 import PageView from './PageView';
 import FixtureRosters from '../components/fixtures/FixtureRosters';
 import FixtureChat from '../components/fixtures/FixtureChat';
@@ -22,9 +17,10 @@ import GameHighlightsCard from '../components/video/GameHighlightsCard';
 import { ProMotmVotingBox } from '../components/pro/motm';
 import ProFixtureVotingBox from '../components/fixtures/voting/ProFixtureVotingBox';
 import { ErrorState } from '../components/ui/ErrorState';
+import FixtureHero from '../components/fixtures/FixtureHero';
 
 export default function FixtureScreen() {
-  const navigate = useNavigate();
+  
   const { fixtureId } = useParams();
 
   const fixtureKey = fixtureId ? `fixture/${fixtureId}` : null;
@@ -89,50 +85,7 @@ export default function FixtureScreen() {
 
   return (
     <div className="dark:text-white flex flex-col">
-      <BlueGradientCard className="p-4 w-full rounded-none h-56 bg-gradient-to-br lg:px-[15%] ">
-        <div
-          onClick={() => navigate(-1)}
-          className="flex mb-5 lg:px-4 cursor-pointer w-full hover:text-blue-500 flex-row items-center justify-start"
-        >
-          <ArrowLeft />
-          <p>Go Back</p>
-        </div>
-
-        <div className="flex flex-row h-max items-center justify-center w-full ">
-          <div className="flex flex-1 flex-col items-center justify-start gap-3">
-            <TeamLogo
-              className="lg:hidden w-12 h-12 dark:text-slate-200 "
-              url={fixture.team.image_url}
-              teamName={fixture.team.athstat_name}
-            />
-            <TeamLogo
-              className="lg:block hidden w-16 h-16 dark:text-slate-200 "
-              url={fixture.team.image_url}
-              teamName={fixture.team.athstat_name}
-            />
-            <p className="text text-wrap text-center">{fixture.team.athstat_name}</p>
-          </div>
-
-          <div className="flex flex-col flex-1">
-            {gameKickedOff && <MatchResultsInformation fixture={fixture} />}
-            {!gameKickedOff && <KickOffInformation fixture={fixture} />}
-          </div>
-
-          <div className="flex flex-1 flex-col items-center gap-3 justify-end">
-            <TeamLogo
-              className="lg:hidden w-12 h-12 dark:text-slate-200 "
-              url={fixture.opposition_team.image_url}
-              teamName={fixture.opposition_team.athstat_name}
-            />
-            <TeamLogo
-              className="lg:block hidden w-16 h-16 dark:text-slate-200 "
-              url={fixture.opposition_team.image_url}
-              teamName={fixture.opposition_team.athstat_name}
-            />
-            <p className="text text-wrap text-center">{fixture.opposition_team.athstat_name}</p>
-          </div>
-        </div>
-      </BlueGradientCard>
+      <FixtureHero fixture={fixture} />
 
       {/* {boxScore && <FixtureScreenTab />} */}
       <FixtureScreenHeader fixture={fixture} />
@@ -180,56 +133,6 @@ export default function FixtureScreen() {
       <div className="flex flex-col p-4 gap-5">
         {/* Overview Component */}
         {loadingSportsActions && <LoadingState />}
-      </div>
-    </div>
-  );
-}
-
-type Props = {
-  fixture: IFixture;
-};
-
-function KickOffInformation({ fixture }: Props) {
-  const { kickoff_time } = fixture;
-
-  return (
-    <div className="flex flex-1 text-nowrap flex-col dark:text-white text-center items-center justify-center">
-      {kickoff_time && <p className="font-bold">{format(kickoff_time, 'h:mm a')}</p>}
-      {kickoff_time && (
-        <p className="dark:text-slate-300 text-slate-200">{format(kickoff_time, 'dd MMM yyyy')}</p>
-      )}
-    </div>
-  );
-}
-
-function MatchResultsInformation({ fixture }: Props) {
-  const { game_status } = fixture;
-
-  return (
-    <div className="flex justify-center  flex-1 w-full flex-col items-center">
-      <div>
-        {game_status && (
-          <span className="text text-slate-white font-semibold dark:text-slate-100">
-            {summerizeGameStatus(fixture)}
-          </span>
-        )}
-      </div>
-
-      <div className="flex flex-1 flex-row gap-2 items-center justify-between">
-        {/* Home Team Score */}
-
-        <div className="dark:text-white flex-1 text-3xl lg:text-4xl font-bold flex items-center justify-end">
-          <p>{fixture.team_score}</p>
-        </div>
-
-        <div className="flex flex-1 flex-col dark:text-white text-center items-center justify-center">
-          <Minus />
-        </div>
-
-        {/* Away Team Score */}
-        <div className="dark:text-white  text-wrap flex-1 text-3xl lg:text-4xl font-bold flex items-center justify-start">
-          <p>{fixture.opposition_score}</p>
-        </div>
       </div>
     </div>
   );
